+++ packages/connect/src/components/EpicAuthProvider/EpicAuthProvider.tsx
@@ -1,6 +1,6 @@
 'use client'
 
-import { Modal, ModalPrimitive, Spinner } from '@0xsequence/design-system'
+import { DialogPrimitive, Modal, Spinner } from '@0xsequence/design-system'
 import { useEffect, type ReactNode } from 'react'
 import { useConnect, useConnection, useConnectors } from 'wagmi'
 
@@ -55,12 +55,12 @@ export const EpicAuthProvider = ({ children }: { children: ReactNode }) => {
             <div className="w-12 h-12 mb-4" aria-label="Epic Games">
               <EpicLogo />
             </div>
-            <ModalPrimitive.Title asChild>
+            <DialogPrimitive.Title asChild>
               <div className="flex items-center gap-4 mt-4 mb-2 flex-row">
                 <h2 className="text-white text-lg font-semibold text-center w-full">Logging in with Epic Games…</h2>
                 <Spinner />
               </div>
-            </ModalPrimitive.Title>
+            </DialogPrimitive.Title>
           </div>
         </Modal>
       )}