+++ packages/connect/src/components/SocialLink/SocialLink.tsx
@@ -1,4 +1,4 @@
-import { Button, Card, Divider, PINCodeInput, Spinner, Text, TextInput } from '@0xsequence/design-system'
+import { Button, Card, PINCodeInput, Separator, Spinner, Text, TextInput } from '@0xsequence/design-system'
 import { type Account } from '@0xsequence/waas'
 import { GoogleLogin, type CredentialResponse } from '@react-oauth/google'
 import { useEffect, useRef, useState, type SetStateAction } from 'react'
@@ -142,7 +142,11 @@ export function SocialLink() {
                     </Text>
                   )}
                 </div>
-                {a.id !== currentAccount?.id && <Button size="xs" label="Remove" onClick={() => removeAccount(a.id)} />}
+                {a.id !== currentAccount?.id && (
+                  <Button size="xs" onClick={() => removeAccount(a.id)}>
+                    Remove
+                  </Button>
+                )}
               </div>
             ))}
           {error && (
@@ -153,7 +157,7 @@ export function SocialLink() {
         </div>
       </Card>
 
-      <Divider />
+      <Separator className="my-4" />
 
       <div className="flex flex-col gap-2 w-full">
         <Text variant="large" color="text100" fontWeight="bold" className="mb-2">
@@ -187,7 +191,7 @@ export function SocialLink() {
           </Text>
         )}
 
-        <Divider />
+        <Separator className="my-4" />
 
         <div className="mt-2">
           <Text variant="normal" color="text100" fontWeight="bold">
@@ -213,10 +217,11 @@ export function SocialLink() {
                 <Button
                   variant="primary"
                   disabled={code.includes('')}
-                  label="Verify"
                   onClick={() => sendChallengeAnswer(code.join(''))}
                   data-id="verifyButton"
-                />
+                >
+                  Verify
+                </Button>
               )}
             </div>
           </div>
@@ -260,10 +265,11 @@ export function SocialLink() {
                 <Button
                   variant="primary"
                   disabled={!isEmailValid}
-                  label="Continue"
                   onClick={() => initiateEmailAuth(email)}
                   data-id="continueButton"
-                />
+                >
+                  Continue
+                </Button>
               )}
             </div>
           </div>