+++ packages/wallet-widget/src/components/NavigationHeader/content/TokenHeader.tsx
@@ -1,6 +1,6 @@
-import { compareAddress, Image, Skeleton, Text } from '@0xsequence/design-system'
+import { compareAddress, getNativeTokenInfoByChainId } from '@0xsequence/connect'
+import { Image, Skeleton, Text } from '@0xsequence/design-system'
 import { useGetSingleTokenBalance } from '@0xsequence/hooks'
-import { getNativeTokenInfoByChainId } from '@0xsequence/web-sdk-core'
 import { zeroAddress } from 'viem'
 import { useConfig } from 'wagmi'